Liquor (Liquor Tags v1.0.0)

Contains the main functions to initialize and render the tags.

Types

t()

@type t() :: %{
  content: String.t(),
  tags: [Liquor.Tags.Tag.t()],
  fetch: (String.t() -> String.t())
}

Initialized the map with the content, tags and fetch. Fetch must call the URL and return the HTML string to extract the OpenGraph data.

Functions

init(content, tags, fetch)

@spec init(String.t(), [Liquor.Tags.Tag.t()], (String.t() -> String.t())) :: t()

Initializes the Liquor map with content, tags and fetch function.

render(content, tags, fetch)

@spec render(String.t(), [Liquor.Tags.Tag.t()], (String.t() -> String.t())) ::
  String.t()

Initializes, finds and render the tags